[DTensor] Support user-supplied Generator for random ops
If the user provides a generator kwarg to a random op (e.g.
nn.init.uniform_(..., generator=my_generator)), we can still advance
that generator's state in a SPMD-global way so that each local-tensor
gets appropriate values and the generator advances to the same state as
if it had operated on the full tensor.

ghstack-source-id: b93f34f4290e3c8da051df538eaebc2988e818e6
Pull Request resolved: https://github.com/pytorch/pytorch/pull/159933